==> 1.fetch-tranfsermarkt-ids-from-wikipedia.ipynb <==
from bs4 import BeautifulSoup
import pandas as pd
from pprint import pprint
import re
import requests
from tqdm.notebook import tqdm
import time

# The ones which were already downloaded
df_old = pd.read_csv("../output/2022-wikipedia-to-transfermarkt.csv")

df_old.head()

# Request
url = 'https://en.wikipedia.org/wiki/2022_FIFA_World_Cup_squads'
r = requests.get(url)
soup = BeautifulSoup(r.text)

# Find the tables
tables = soup.find_all('table')

# Check if the tables have the wanted columns
tables = [table for table in tables if len(table.find_all("abbr")) == 2] # only player tables have 2 abbr
len(tables)

# For each table, find all the player links
player_links = []
for table in tables:
    players = table.find_all("tr", class_='nat-fs-player')
    
    for player in players:
        link = player.find('th').find('a')['href']
        player_links.append(link)

# After finding all players, we will retrieve the wikidata entry for each one of them
wikibase_list = []
for index, player_url in enumerate(tqdm(player_links)):
    
    query = f"https://en.wikipedia.org{player_url}"
    
    if query in df_old.wikipedia_page.unique().tolist():
        continue
    
    print(query)
    
    r = requests.get(query)
    soup = BeautifulSoup(r.text)
        
    wikibase_url = soup.find("li", {"id": "t-wikibase"}).find('a')['href']
    wikibase_item = re.search("\d+", wikibase_url).group(0)
    

    datapoint = {"query": query, 
                     "wikibase_url": wikibase_url, 
                     "wikibase_item": wikibase_item}
                              
    print(datapoint)
    
    wikibase_list.append(datapoint)

dataset = []
scrapped = []

# For each player,get their transfermarkt id from wikidata
for wikibase_item in tqdm(wikibase_list):
    
    wikibase_url = wikibase_item['wikibase_url']
    wikipedia_page = wikibase_item['query']
    wikibase_id = wikibase_item['wikibase_item']
    
    if wikipedia_page in df_old.wikipedia_page.unique().tolist():
        continue
        
    print(wikibase_id)
    
    r = requests.get(wikibase_url)
    soup = BeautifulSoup(r.text)

    div = soup.find("div", {"data-property-id": "P2446"})
    transfermarkt_id = div.find("a", class_="wb-external-id external")

    try:
        transfermarkt_name = div.find_all("div", class_="wikibase-snakview-value wikibase-snakview-variation-valuesnak")[1].text
    except IndexError:
        transfermarkt_name = 'not on wikidata'

    row = {
        'wikipedia_page': wikipedia_page,
        'transfermarkt_id': transfermarkt_id.text,
        'transfermarkt_name': transfermarkt_name
    }
    
    print(row)
    
    dataset.append(row)
    scrapped.append(wikibase_id)

df = pd.DataFrame(dataset)

df = pd.concat([df_old, df])

df.to_csv("../output/2022-wikipedia-to-transfermarkt.csv", index=False)
