==> 2.fetch-player-data.ipynb <==
'''
This scripts collects information about
each player. It collects his basic information,
his transfer history and all the leagues he played for.
'''

from bs4 import BeautifulSoup
from io import BytesIO
import glob
import json
import pandas as pd
from pathlib import Path
from pprint import pprint
import requests
import re
from tqdm.notebook import tqdm
import time

HEADERS = {
        'User-Agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/47.0.2526.106 Safari/537.36'
    }

def already_saved(player_id):
    '''
    Controls whether the data for this specific player
    was already saved properly. Returns true or false.
    
    Params:
    
    player_id -> string with the unique player id at Transfermarkt
    '''
    
    player_files = glob.glob("../output/players/basic-info-*.json")
    transfer_files = glob.glob("../output/players/transfer-info-*.json")
    league_files = glob.glob("../output/players/league-info-*.json")
    club_files = glob.glob("../output/players/club-info-*.json")
    
    player_files = list(map(lambda x: re.search("\d+", x).group(0), player_files))
    transfer_files = list(map(lambda x: re.search("\d+", x).group(0), transfer_files))
    league_files = list(map(lambda x: re.search("\d+", x).group(0), league_files))
    club_files = list(map(lambda x: re.search("\d+", x).group(0), club_files))
    
    if player_id in player_files and player_id in league_files and \
    player_id in club_files and player_id in transfer_files:
        
        return True

def player_page_request(player_url): 
    '''
    Makes a request to the page containing
    detailed information about the player
    and returns it as a response object.
    
    Params:
    
    player_url -> string with the url path to the player page at Transfermarkt
    '''
    
    r = requests.get(player_url, headers=HEADERS)
    
    return r

def scrape_basic_info(player_r):
    '''
    Collects the basic player information.
    
    Params:
    
    player_r -> a response object retrieved from 
    the player page at transfermarkt
    '''
    
    player_info = {}
    
    soup = BeautifulSoup(player_r.text)

    player_data = soup.find('div', class_='info-table')
    categories = player_data.find_all('span', class_='info-table__content--regular')
    values = player_data.find_all('span', class_='info-table__content--bold')

    for cat, val in zip(categories, values):
        
        if cat.text.strip() in ['Name in home country:', 'Full name:', 'Date of birth:', 
                        'Age:', 'Height:', 
                        'Position:', 'Foot:', 'Current club:', 'Joined:']:

            label = cat.text.replace(":", "").strip().lower().replace(" ", "_")
            item = val.text.strip()

            player_info[label] = item
            
        if cat.text.strip() in ['Current club:']:
            player_info['current_club_url'] = val.find('a', href=True)['href']
            player_info['current_club_id'] = re.search("/verein/(\d+)", player_info['current_club_url']).group(1)


        if cat.text.strip() == "Citizenship:":
            citizenship = val.find("img", class_="flaggenrahmen", alt=True)['alt'] # In cases of double citizenship, this will get the first one only
            player_info['citizenship'] = citizenship
            
        if cat.text.strip() == 'Place of birth:':

            # Add country before saving
            label = cat.text.replace(":", "").strip().lower().replace(" ", "_")

            country = val.find('img', class_='flaggenrahmen', alt=True)['alt']
            item = f'{val.text.strip()}, {country}'

            player_info[label] = item
            
    return player_info

def save_json(data, outpath):
    '''
    Saves a JSON file in the specified output path
    '''
    with open(outpath, 'w+') as f:
        json.dump(data, f)

def scrape_transfer_info(player_r):
    '''
    Scrapes all transfers from a player
    
    Params:
    
    player_r -> a response object retrieved from 
    the player page at transfermarkt
    '''
    
    soup = BeautifulSoup(player_r.text)
    transfers = soup.find_all('div', class_='grid tm-player-transfer-history-grid')
    
    rows = []

    for transfer in transfers:
        entries = transfer.find_all("div", class_="grid__cell")
        # Removes the fee and market value entries
        entries = entries[:-2]

        season = entries[0].text.strip()
        date = entries[1].text.strip()
        left = entries[2].text.strip()
        left_url = entries[2].find('a')['href']
        left_club_id = re.search('verein/(\d+)/saison_id/', left_url).group(1)
        joined = entries[3].text.strip()
        joined_url = entries[3].find('a')['href']
        joined_club_id = re.search('verein/(\d+)/saison_id/', joined_url).group(1)

        datum = {
            'season': season,
            'date': date,
            'left': left,
            'left_url': left_url,
            'left_club_id': left_club_id,
            'joined': joined,
            'joined_url': joined_url,
            'joined_club_id': joined_club_id,

        }

        rows.append(datum)
        
    return rows

def scrape_league_info(player_id):
    '''
    Scrapes the league information for each player.

    Params:

    player_id -> string with the unique player id at Transfermarkt
    '''
    
    # Funny, we don't actually need the player name on the url
    league_details_url = f'https://www.transfermarkt.com/dummy/leistungsdaten/spieler/{player_id}/plus/0?saison=ges'
    print(league_details_url)
    
    time.sleep(1)

    r = requests.get(league_details_url, headers=HEADERS)
    soup = BeautifulSoup(r.text)
    
    
    league_table_div = soup.find('div', class_='responsive-table')
    league_table = league_table_div.find("table", class_='items')
    tbody = league_table.find("tbody")
    
    rows = tbody.find_all("tr")
    
    data = []
    for row in rows:
        cells = row.find_all("td")

        league = cells[1].find('a')
        league_url = league['href']
        league_name = league['title']
        league_id = re.search("\/startseite\/\w+\/(.+)", league_url).group(1)

        matches = cells[2].text
        minutes = cells[-1].text

        datum = {
            "league_url": league_url,
            "league_name": league_name,
            "league_id": league_id,
            "matches": matches,
            "minutes": minutes,
        }
        
        data.append(datum)
                  

    return data


def scrape_club_info(player_id):
    '''
    Scrapes the club information for each player.
    
    Params:

    player_id -> string with the unique player id at Transfermarkt
    '''
    
    club_details_url = f'https://www.transfermarkt.com/dummy/leistungsdatenverein/spieler/{player_id}'
    print(club_details_url)
    
    time.sleep(1)

    
    r = requests.get(club_details_url, headers=HEADERS)
    soup = BeautifulSoup(r.text)

    club_table_div = soup.find('div', class_='responsive-table')
    club_table = club_table_div.find("table", class_='items')
    
    tbody = club_table.find("tbody")
    
    rows = tbody.find_all("tr")
    
    data = []
    for row in rows:
        cells = row.find_all("td")

        club = cells[1].find('a')
        club_url = club['href']
        club_name = club['title']
        club_id = re.search("\/startseite\/\w+\/(.+)", club_url).group(1)

        matches = cells[2].text
        minutes = cells[-1].text

        datum = {
            "club_url": club_url,
            "club_name": club_name,
            "club_id": club_id,
            "matches": matches,
            "minutes": minutes,
        }
        
        data.append(datum)
                  

    return data

def main():
    
    df = pd.read_csv('../output/2022-wikipedia-to-transfermarkt.csv', sep=',')
    
    all_files = glob.glob("../output/players/*.json")
    right_ids = df.transfermarkt_id.astype(str).unique().tolist()
        
    for file in all_files:
        
        id_ = re.search("(\d+)", file).group(1)
        
        if id_ not in right_ids:
            print('deleting', file) # to do – turn into delete
            file_to_rem = Path(file)
            file_to_rem.unlink()
    
    for index, row in tqdm(df.iterrows(), total=df.shape[0]):
                
        print(row.wikipedia_page)
        
        # Progress check
        player_url = f"https://www.transfermarkt.com/dummy/profil/spieler/{row.transfermarkt_id}"
        player_id = re.search('/spieler/(\d+)', player_url).group(1)

        if already_saved(player_id):
            print("Already collected!")
            print()
            continue
            

        # If the players was not saved, let's fetch basic data
        print(player_url)
        player_r = player_page_request(player_url)
        
        time.sleep(1)

        player_info = scrape_basic_info(player_r)
#         player_info['fetched_name'] = row.fetched_name
#         player_info['national_team'] = row.team
        save_json(player_info, f"../output/players/basic-info-{player_id}.json")
        
        # Scrapes transfer data
        transfer_info = scrape_transfer_info(player_r)
        save_json(transfer_info, f"../output/players/transfer-info-{player_id}.json")

        # Collect league data
        league_info = scrape_league_info(player_id)
        save_json(league_info, f"../output/players/league-info-{player_id}.json")

          # Collect club data
        club_info = scrape_club_info(player_id)
        save_json(club_info, f"../output/players/club-info-{player_id}.json")

        print()

if __name__ == "__main__":
    main()
