==> 4.fetch-league-info.ipynb <==
'''
Fetches league information for each of the competitions
and saves as a single CSV, also updating the existing JSON files
in the process.
'''

from bs4 import BeautifulSoup
from io import BytesIO
import glob
import json
import pandas as pd
from pprint import pprint
import requests
import re
from selenium import webdriver
from tqdm.notebook import tqdm
import time

HEADERS = {
        'User-Agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/47.0.2526.106 Safari/537.36'
    }

def already_saved(league_id):
    '''
    Controls whether the data for this specific league
    was already saved properly. Returns true or false.
    
    Params:
    
    club_id -> string with the unique club id at Transfermarkt
    '''
    
        
    leagues = glob.glob("../output/leagues/*.json")
    leagues = list(map(lambda x: re.search("/leagues/(.+).json", x).group(1), leagues))

    return (league_id in leagues)

def fetch_leagues(pattern):
    '''
    Fetches all files with league information
    and keeps unique entries to scrape.
    
    Params:
    
    pattern -> str, a glob pattern
    '''
    
    league_files = glob.glob(pattern)
    
    dfs = []
    for file in league_files:
        with open(file, "r") as f:
            data = json.load(f)
            
        df = pd.DataFrame(data)
        dfs.append(df)
        
    dfs = pd.concat(dfs, ignore_index=True)
    
    dfs = dfs[['league_url', 'league_name', 'league_id']]
    dfs = dfs.drop_duplicates().reset_index(drop=True)
    
    return dfs

def save_json(data, outpath):
    '''
    Saves a JSON file in the specified output path
    '''
    with open(outpath, 'w+') as f:
        json.dump(data, f)

def get_league_info(league_url):
    '''
    Scrapes the relevant club information.
    
    Params:
    
    driver -> a selenium webdriver
    club_url -> part of the url for the transfermakt club page
    '''
    
    url = f"https://transfermarkt.com{league_url}"
    print(url)
    
    r = requests.get(url, headers=HEADERS)
    
    soup = BeautifulSoup(r.text)
    
    data = {}
    
    # Cup
    if 'pokalwettbewerb' in league_url:
        content = soup.find('div', class_='dataContent')
        data_value = content.find('span', class_='dataValue')
    
        country = data_value.find('img')
        if not country: 
            country = 'no country'
        else:
            country = country['alt'].strip()
            
        league_tier = data_value.text.strip()
        
        data['country'] = country
        data['league_tier'] = league_tier
        data['type'] = 'cup'
        
    # League
    elif 'wettbewerb' in league_url:
        
        box = soup.find("div", class_="data-header__box--big")
        country = box.find("span", class_="data-header__club").find('a').text.strip()
        league_tier = box.find("span", class_="data-header__label").find('span').text.strip()
        
        data['country'] = country
        data['league_tier'] = league_tier
        data['type'] = 'league'

   
    return data

def main():
    
    leagues_to_scrape = fetch_leagues("../output/players/league-info-*.json")
        
    for index, row in tqdm(leagues_to_scrape.iterrows(), total=leagues_to_scrape.shape[0]):
  
        league_url = row.league_url
        league_id = row.league_id
        league_name = row.league_name
        
        print(f"{league_id}, {league_name}")
        
        if already_saved(league_id):
            print('Already saved')
            print()
            continue
    
        data = get_league_info(league_url)
        data['league_url'] = league_url
        data['league_id'] = league_id
        data['league_name']= league_name
        
        pprint(data)

        with open(f"../output/leagues/{league_id}.json", "w+") as f:
            json.dump(data, f)
            
        print()

if __name__ == "__main__":
    main()
