==> 7.files-for-dataviz.ipynb <==
import pandas as pd
import json
import numpy as np

#### Interactive?

df = pd.read_csv("../output/2022-full-player-data.csv")

names_and_links = pd.read_csv("../output/2022-wikipedia-to-transfermarkt.csv")

df = df.merge(names_and_links, left_on='player_id', right_on='transfermarkt_id')

df['share_days_abroad'] = (df.days_abroad / df.total_days).round(2)
df['share_days_home'] = (df.days_home / df.total_days).round(2)

df = df[["player_id", "transfermarkt_name", "date_of_birth", "age", "place_of_birth", "share_days_abroad", "share_days_home", "citizenship", "current_club"]]

for country in df.citizenship.unique():
    
    fname = country.lower().replace(" ","-")
    
    # Saves CSV
    df[df.citizenship==country].to_csv(f"../viz/data/teams/{fname}.csv", index=False)
    
    
    data = {
        "country": country,
        "share_days_home": df[df.citizenship==country].share_days_home.describe().round(2).to_dict(),
        "share_days_abroad": df[df.citizenship==country].share_days_home.describe().round(2).to_dict()
    }
    # And a json with sumarry statistics
    fpath = f"../viz/data/summaries/{fname}.json"
    
    with open(fpath, "w+") as file:
        json.dump(data, file)

df[df.citizenship==country].to_csv(f"../viz/data/teams/{country}.csv", index=False)

df[df.citizenship==country].to_csv(f"../viz/data/teams/{country}.csv", index=False)

df.share_days_home.describe().round(2).to_dict()

selected_countries = [
    "England", "Germany", "Spain", #"France",
    "Brazil", "Argentina", "Ecuador", "Uruguay",
    "Cameroon", "Morocco", "Senegal", "Ghana", "Tunisia"
]

### By groups of interest
caf_teams = ['Ghana', 'Senegal', 'Morocco', 'Cameroon', "Tunisia"]
conmebol_teams = ['Brazil', 'Argentina', 'Uruguay', 'Ecuador']
concacaf_teams = ['United States', 'Mexico', 'Canada', 'Costa Rica']
afc_teams = ['Australia', 'Iran', 'Japan', 'Qatar', 'Saudi Arabia', 'Korea, South']
uefa_teams = [ 'Belgium', 'Croatia', 'Denmark', 'England', 'France', 'Germany',
              'Netherlands', 'Poland', 'Portugal', 'Serbia', 'Spain', 'Switzerland',
              'Wales']

def continent(row):
    
    if row.citizenship in caf_teams:
        return 'africa'
    elif row.citizenship in conmebol_teams:
        return 'south-america'
    elif row.citizenship in afc_teams:
        return 'asia'
    elif row.citizenship in concacaf_teams:
        return 'central-and-north-america'
    elif row.citizenship in uefa_teams:
        return 'europe'

df['continent'] = df.apply(continent, axis=1)

df[df.citizenship.isin(selected_countries)].to_csv("../viz/data/selected-countries.csv", index=False)

df.to_csv("../viz/data/all-countries.csv", index=False)

#### Dotplot

df = pd.read_csv("../output/2022-full-player-data.csv")

names_and_links = pd.read_csv("../output/2022-wikipedia-to-transfermarkt.csv")

df = df.merge(names_and_links, left_on='player_id', right_on='transfermarkt_id')
